Add integration tests for appointment status being recalculated after creating availability
